feat(SearchOverlay): implement search overlay components and hooks for improved search functionality
